arch: x3: MTK: AEE: Disable MTK_AEE
DO NOT REVERT unless you plan on fixing the issue.

Disabled because its just a debugging module, and its not "required".
Main reason for removing it is that it uses android logger, which is removed by google now and is not present in the kernel.
It is not modified to work without the android logger, so we just removed it.

Signed-off-by: Sahil Soni <SscSPs@gmail.com>